Utilities header added. Floor Bombs now generate tiles.
Utilities header is used for various project-wide functions, like spawning a Blueprint Actor from code.

Work on Floor Bombs continued. Upon hitting an object in the world or the second click of the "Bomb" button, the Floor Bomb now detonates and creates floor tiles in that area.

Various tweaks and cleanup performed.