Added moveable tiles
Initial implementation of moveable tiles.
A moveable tile has an array of actors that are considered "checkpoints" and moves between them.

The concept will be expanded upon with different types of moveable tiles, as well as stronger checkpoint constraints (i.e. currently, any actors can be checkpoints, but they should be actors of a specific subclass)